sys-kernel/coreos-modules: Install external mod build files with script
The kernel now includes a script for installing the files needed to
build out-of-tree modules, rendering our existing code obsolete. The
layout is different, but we were following Ubuntu's non-standard layout
when there was no need to. Ubuntu's approach is seemingly designed to
save space by symlinking common files across different platforms, but
Flatcar doesn't need to do this.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>