overlay profiles: Unmask thin-provisioning-tools, add accept keywords for it
No point in masking the new versions of the package - only those are
available now, so the mask was preventing building any available
version of the package.

At the same time, we need to pull an unstable version 1.3.0, as our
rust version is newer than allowed by the older available 1.1.0
version.

Signed-off-by: Krzesimir Nowak <knowak@microsoft.com>